requirements.txt
django==2.1.1
djangorestframework==3.8.2
markdown==3.0.1
coverage==4.5.1
coreapi==2.3.3